% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{download_tools}
\alias{download_tools}
\title{Download tools for offline projects}
\usage{
download_tools(tools = c("R", "Rtools", "RStudio"))
}
\arguments{
\item{tools}{A string vector specifying tools}
}
\description{
Download tools for offline projects
}
