% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install_packages}
\alias{install_packages}
\alias{install_project_packages}
\title{Install packages at once}
\usage{
install_packages(packages)

install_project_packages()
}
\arguments{
\item{packages}{A string vector specifying packages}
}
\description{
Install packages at once
}
