% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offline-package.R
\docType{package}
\name{offline-package}
\alias{offline}
\alias{offline-package}
\title{offline: Preparation of offline proejcts}
\description{
This package is for personal purposes to prepare offline projects.
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
