% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_r_packages}
\alias{zip_r_packages}
\title{Create installed R packages zip archive}
\usage{
zip_r_packages(flags = "-r9X", split_size)
}
\arguments{
\item{flags}{A string of flags to be passed to the command}

\item{split_size}{A numeric value of split size (Megabyte)}
}
\description{
Create installed R packages zip archive
}
\details{
flag options
\itemize{
\item -r  recurse into directories
\item -0  store only
\item -1  compress faster
\item -9  compress better
\item -X  eXclude eXtra file attributes
\item -q  quiet operation
\item -s  split size
}
}
